bump version of mdadm
  Thanks to Ivan Kuten.
  Minor cleanup while at it (me).

+++ package/mdadm/mdadm.mk
@@ -5,7 +5,7 @@
 #############################################################
 MDADM_VERSION:=2.6.2
 MDADM_SOURCE:=mdadm_$(MDADM_VERSION).orig.tar.gz
-MDADM_PATCH:=mdadm_$(MDADM_VERSION)-1.diff.gz
+MDADM_PATCH:=mdadm_$(MDADM_VERSION)-2.diff.gz
 MDADM_CAT:=$(ZCAT)
 MDADM_SITE:=http://ftp.debian.org/debian/pool/main/m/mdadm
 MDADM_DIR:=$(BUILD_DIR)/mdadm-$(MDADM_VERSION)
@@ -39,7 +39,9 @@ $(MDADM_DIR)/$(MDADM_BINARY): $(MDADM_DIR)/.unpacked
 
 $(TARGET_DIR)/$(MDADM_TARGET_BINARY): $(MDADM_DIR)/$(MDADM_BINARY)
 	$(MAKE) DESTDIR=$(TARGET_DIR) -C $(MDADM_DIR) install
+ifneq ($(BR2_HAVE_MANPAGES),y)
 	rm -Rf $(TARGET_DIR)/usr/share/man
+endif
 	$(STRIPCMD) $(STRIP_STRIP_ALL) $@
 
 mdadm: uclibc $(TARGET_DIR)/$(MDADM_TARGET_BINARY)
